erscom.rs: Fix password visibility resetting after every keystroke in the settings dialog
This is because setting back to the same variable we got our values from
caused a react-style recreation of the widget which caused the visibility to
go back to the default.

Since I'm not exactly sure about the granularity of the data dependencies
that slint uses, I just moved the state back to the rust code and added a
set() callback. The save() callback now has no parameters.

This actually simplifies the rust a little bit.